+++ doc/source/changelog.rst
@@ -9,6 +9,8 @@ New features
 
 * Database size monitoring - new panel (in admin menu)
 * Sofia Gateway : add new chooce lists for selecting codecs
+* Add sip profile DTMF options (pass-rfc2833)
+* Add new rtp_rewrite_timestamps sofia profile option
 
 Modifications
 -------------